LPD-96652 Move BaseCard and SectionHeader to common
