+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/inventory/components/ExpiredAssetsCard.tsx
@@ -15,6 +15,7 @@ import React, {useContext, useEffect, useState} from 'react';
 
 import ApiHelper from '../../../../common/services/ApiHelper';
 import {openCMSModal} from '../../../../common/utils/openCMSModal';
+import {BaseCard} from '../../common/BaseCard';
 import {InventoryContext} from '../InventoryContext';
 import {
 	AssetType,
@@ -24,7 +25,6 @@ import {
 	AssetTypeIcons as _AssetTypeIcons,
 } from '../utils/assetTypes';
 import usePagination from '../utils/usePagination';
-import {BaseCard} from './BaseCard';
 import {Item} from './FilterDropdown';
 
 type ExpiredAsset = {
