+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/inventory/components/FilesCard.tsx
@@ -11,7 +11,7 @@ import {
 import {sub} from 'frontend-js-web';
 import React, {useState} from 'react';
 
-import {BaseCard} from './BaseCard';
+import {BaseCard} from '../../common/BaseCard';
 import {ContentAndFilesCard} from './ContentAndFilesCard';
 
 export function FilesCard() {
